requirements.txt
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
oauthlib==3.1.0
prometheus-client==0.8.0
psycopg2-binary==2.8.6
python-twitter==3.5
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
SQLAlchemy==1.3.19
urllib3==1.25.10
